+++ README.md
@@ -66,6 +66,8 @@ del: 1,000,000 ops over 4 threads in 25292ms, 39,538/sec, 25291 ns/op, 42.2 MB, 
 
 ```
 
+
+
 ## How it is done
 
 * Sniper database is sharded on many chunks. Each chunk has its own lock (RW), so it supports high concurrent access on multi-core CPUs.
