+++ bench/go.mod
@@ -1,0 +1,8 @@
+module sniperbench
+
+go 1.13
+
+require (
+	github.com/recoilme/sniper@dev
+	github.com/tidwall/lotsa v0.0.0-20190911211615-e96c7fea75f1
+)
